Code refactoring. Add AuthViewModel class as mediator between presentation and business logic. Moved around some packages so it looks more structured
