+++ client/app/src/main/java/ru/hse/online/client/repository/networking/api_service/AuthApiService.kt
@@ -1,9 +1,9 @@
-package ru.hse.online.client.networking.api_service
+package ru.hse.online.client.repository.networking.api_service
 
 import retrofit2.http.Body
 import retrofit2.http.POST
-import ru.hse.online.client.networking.api_data.AuthRequest
-import ru.hse.online.client.networking.api_data.AuthResponse
+import ru.hse.online.client.repository.networking.api_data.AuthRequest
+import ru.hse.online.client.repository.networking.api_data.AuthResponse
 
 interface AuthApiService {
     @POST("/login")
