+++ client/app/src/main/java/ru/hse/online/client/view/map/MapView.kt
@@ -1,4 +1,4 @@
-﻿package ru.hse.online.client.view
+﻿package ru.hse.online.client.view.map
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -7,15 +7,12 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import ru.hse.online.client.view.map.ActualMapView
-import ru.hse.online.client.view.map.BaseMapView
-import ru.hse.online.client.view.map.MapOverlayView
 import ru.hse.online.client.ui.theme.ClientTheme
 
 class MapView : ComponentActivity() {
 //    private val mapView: BaseMapView = StaticMapView()
-    private val mapView: BaseMapView = ActualMapView()
-    private val mapOverlay: MapOverlayView = MapOverlayView()
+    private val mapView: BaseMapView = GoogleMapView()
+    private val mapOverlay: MapOverlayView = MapOverlayView(this)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -25,7 +22,7 @@ class MapView : ComponentActivity() {
             ClientTheme(darkTheme = true) {
                 Surface(modifier = Modifier.fillMaxSize()) {
                     mapView.DrawMap()
-                    mapOverlay.Draw(this)
+                    mapOverlay.Draw()
                 }
             }
         }
